pyproject.toml
[tool.poetry]
name = "nzshm-hazlab"
version = "0.1.0"
description = "A laboratory for exploring hazard models"
authors = ["chrisdicaprio <christopher.dicaprio@gmail.com>"]
license = "AGPL-3.0"
readme = "README.md"
packages = [{include = "nzshm_hazlab"}]

[tool.poetry.dependencies]
python = "^3.8"
matplotlib = "^3.6.2"
ipython = {version = "^8.7.0", optional = true}
pyqt5 = {version = "^5.15.7", optional = true}
markdown = "^3.4.1"
pygmt = "^0.8.0"
notebook = {version = "^6.5.3", optional = true}
geopandas = "^0.12.2"
toml = "^0.10.2"
toshi-hazard-store = "^0.7.2"
nzshm-runzi = {git = "https://github.com/GNS-Science/nzshm-runzi.git", rev = "feature/148-disagg-error-handling", optional = true}
xmltodict = "^0.13.0"
toshi-hazard-post = {path = "/home/chrisdc/NSHM/DEV/CALCULATION/toshi-hazard-post", optional = true}

[tool.poetry.extras]
ipython = ["ipython", "pyqt5"]
runzi = ["nzshm-runzi"]
jupyter = ["notebook"]
openquake = ["openquake-engine"]
thp = ["toshi-hazard-post"]

[tool.poetry.group.dev.dependencies]
mypy = "^0.991"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
